requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
PySimpleSOAP==1.16.2
pytz==2021.3
sqlparse==0.4.2
suds==1.0.0
tzdata==2021.5
